requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.23.0
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
sqlparse==0.4.2
threadpoolctl==3.1.0
